Clean/Split up mass configuration options
